No ListCellRenderer for non-lists
